remove unused files and update toast on android
