+++ src/view/com/composer/Composer.tsx
@@ -117,7 +117,6 @@ import {Button, ButtonIcon, ButtonText} from '#/components/Button'
 import {CircleInfo_Stroke2_Corner0_Rounded as CircleInfo} from '#/components/icons/CircleInfo'
 import {EmojiArc_Stroke2_Corner0_Rounded as EmojiSmile} from '#/components/icons/Emoji'
 import {LazyQuoteEmbed} from '#/components/Post/Embed/LazyQuoteEmbed'
-import {SelectVideoBtn} from '#/view/com/composer/videos/SelectVideoBtn'
 import {SubtitleDialogBtn} from '#/view/com/composer/videos/SubtitleDialog'
 import {Text as NewText} from '#/components/Typography'
 import {Text} from '#/view/com/util/text/Text'
